refactor: extract direct thinking effort policy
